refactor: Account 엔티티 분리
- Account 엔티티를 CheckingAccount(메인계좌)와 SavingAccount(적금계좌)로 분리하고자 합니다.
- 그에 따라 Account 엔티티의 필드에서 enum 클래스인 AccountType(MAIN, SAVING)을 제거하였습니다.
- AccountType enum 클래스는 CHECKING, SAVING 값으로 변경하였고 AccountNumberCreator 클래스가 계좌번호를 생성할 때, 파라미터로 넘겨줍니다.
