+++ src/main/java/com/jindo/minipay/account/common/type/AccountType.java
@@ -1,0 +1,6 @@
+package com.jindo.minipay.account.common.type;
+
+public enum AccountType {
+  CHECKING, SAVING
+
+}
