--- src/main/java/com/jindo/minipay/account/constant/AccountConstants.java
@@ -1,9 +1,0 @@
-package com.jindo.minipay.account.constant;
-
-public class AccountConstants {
-
-  public static final String MAIN_ACCOUNT_PREFIX = "1111";
-
-  public static final String SAVINGS_ACCOUNT_PREFIX = "1122";
-
-}
