--- src/main/java/com/jindo/minipay/account/entity/type/AccountType.java
@@ -1,10 +1,0 @@
-package com.jindo.minipay.account.entity.type;
-
-/**
- * MAIN : 메인 계좌
- * SAVINGS : 적금 계좌
- */
-public enum AccountType {
-  MAIN, SAVINGS
-  ;
-}
